chore: refactor existing hook, clean up test cases
