+++ src/hooks/useDailyStatistics/utils.test.ts
