+++ src/hooks/useDailyStatistics/utils.ts
