--- src/hooks/useFetchDailyStatistics/useFetchDailyStatistics.test.tsx
@@ -1,136 +1,0 @@
-import React, { FunctionComponent } from "react";
-import { renderHook } from "@testing-library/react-hooks";
-import { useDailyStatistics } from "./useFetchDailyStatistics";
-import { getDailyStatistics } from "../../services/statistics";
-import { ProductContextProvider } from "../../context/products";
-import { defaultIdentifier } from "../../test/helpers/defaults";
-import { CampaignPolicy } from "../../types";
-
-jest.mock("../../services/statistics");
-const mockGetDailyStatistics = getDailyStatistics as jest.Mock;
-
-const key = "KEY";
-const endpoint = "https://myendpoint.com";
-const operatorToken = "operator-token";
-
-const wrapper: FunctionComponent = ({ children }) => (
-  <ProductContextProvider products={customProducts}>
-    {children}
-  </ProductContextProvider>
-);
-
-const customProducts: CampaignPolicy[] = [
-  {
-    category: "tt-token",
-    categoryType: "DEFAULT", // 'DEFAULT' (default) | 'APPEAL'
-    name: "TT Token",
-    order: 1,
-    identifiers: [
-      {
-        ...defaultIdentifier,
-        label: "first"
-      }
-    ],
-    quantity: {
-      period: -1,
-      periodType: "ROLLING",
-      periodExpression: 365,
-      limit: 1,
-      default: 1
-    },
-    type: "REDEEM"
-  },
-  {
-    category: "tt-token-lost",
-    categoryType: "APPEAL",
-    name: "Lost/stolen token",
-    order: 1,
-    alert: {
-      threshold: 2,
-      label: "*chargeable"
-    },
-    quantity: {
-      period: -1,
-      periodType: "ROLLING", // ROLLING | CRON;
-      periodExpression: 365, // TBD
-      limit: 9999,
-      default: 1
-    },
-    identifiers: [
-      {
-        ...defaultIdentifier,
-        label: "first"
-      }
-    ],
-    type: "REDEEM"
-  }
-];
-
-describe("useFetchDailyStatistics", () => {
-  beforeEach(() => {
-    jest.resetAllMocks();
-  });
-
-  it("daily statistics hook should be initialised correctly", async () => {
-    expect.assertions(4);
-    const { result } = renderHook(() =>
-      useDailyStatistics(key, endpoint, operatorToken, Date.now())
-    );
-    expect(result.current.error).toBeUndefined();
-    expect(result.current.totalCount).toBeNull();
-    expect(result.current.lastTransactionTime).toBeNull();
-    expect(result.current.transactionHistory).toMatchObject([]);
-  });
-
-  it("should set error if error is thrown while getting daily stats", async () => {
-    expect.assertions(1);
-
-    mockGetDailyStatistics.mockRejectedValue(
-      new Error("Error getting daily statistics")
-    );
-
-    const { result, waitForNextUpdate } = renderHook(
-      () => useDailyStatistics(key, endpoint, operatorToken, Date.now()),
-      { wrapper }
-    );
-
-    await waitForNextUpdate();
-
-    expect(result.current.error?.message).toBe(
-      "Error getting daily statistics"
-    );
-  });
-
-  it("should set last transaction time, total count, transaction history after getting daily stats", async () => {
-    expect.assertions(3);
-
-    mockGetDailyStatistics.mockResolvedValue({
-      pastTransactions: [
-        {
-          category: "vouchers",
-          quantity: 9999999,
-          transactionTime: new Date(12000000000)
-        }
-      ]
-    });
-
-    const { result, waitForNextUpdate } = renderHook(
-      () => useDailyStatistics(key, endpoint, operatorToken, 12000000000),
-      { wrapper }
-    );
-
-    await waitForNextUpdate();
-
-    expect(result.current.lastTransactionTime).toStrictEqual(
-      new Date(12000000000)
-    );
-    expect(result.current.totalCount).toBe(9999999);
-    expect(result.current.transactionHistory).toStrictEqual([
-      {
-        category: "vouchers",
-        name: "vouchers",
-        quantityText: "9,999,999 qty"
-      }
-    ]);
-  });
-});
